refactor: Auctions section
